refactor run commands to separate file
